Initial commit
This commit adds the initial index.html, data, and javascript files for the
explorer project. This project will include images take from various stages in
the computer vision process as stacks that can be traversed. It will also
include a three dimensional model of the chromosomes that can zoomed and
rotated.